requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
colorama==0.4.6
distro==1.9.0
Django==5.1.5
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.2.2
pandas==2.2.3
pip==24.3.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
wheel==0.45.1
 cohere==5.13.11

